# Improve load times for project and builds page