--- a/squad/frontend/templates/squad/_builds_table.jinja2
+++ b/squad/frontend/templates/squad/_builds_table.jinja2
@@ -4,8 +4,8 @@
   <input type="hidden" ng-init="invalid_number_of_builds_msg='{{_('Two builds need to be selected!')}}'" />
   <input type="hidden" ng-init="invalid_selected_builds_msg='{{_('Selected builds must be different!')}}'" />
   <input type="button" ng-value="compare_button_label" class="btn btn-default" ng-click="compareBuilds('{{project.full_name}}')" />
-{% for status in statuses %}
-{% with build=status.build %}
+{% for build in builds %}
+{% with status=build.status %}
 <div class="row row-bordered build">
     <div class="col-md-1 col-sm-1">
         <input type="radio" name="baseline_build" ng-value="'{{build.version}}'" ng-model="baseline_build" />&nbsp;