--- a/squad/frontend/templates/squad/builds.jinja2
+++ b/squad/frontend/templates/squad/builds.jinja2
@@ -7,7 +7,7 @@
 {% endwith %}
 <div style='margin-top: 20px'></div>
 
-{% with items=statuses %}
+{% with items=builds %}
   {% include "squad/_pagination.jinja2" %}
 {% endwith %}
 
@@ -15,7 +15,7 @@
 {% include "squad/_builds_table.jinja2" %}
 </div>
 
-{% with items=statuses %}
+{% with items=builds %}
   {% include "squad/_pagination.jinja2" %}
 {% endwith %}